+++ app/models/pay/subscription.rb
@@ -35,7 +35,7 @@ module Pay
     attribute :prorate, :boolean, default: true
 
     # Helpers for payment processors
-    %w[braintree stripe paddle].each do |processor_name|
+    %w[braintree stripe paddle fake_processor].each do |processor_name|
       define_method "#{processor_name}?" do
         processor == processor_name
       end
@@ -92,6 +92,11 @@ module Pay
 
     def has_incomplete_payment?
       past_due? || incomplete?
+    end
+
+    def change_quantity(quantity)
+      payment_processor.change_quantity(quantity)
+      update(quantity: quantity)
     end
 
     def resume