+++ lib/generators/active_record/templates/migration.rb
@@ -2,7 +2,7 @@
 
 class AddPayBillableTo<%= table_name.camelize %> < ActiveRecord::Migration<%= migration_version %>
   def change
-    change_table :<%= table_name %> do |t|
+    change_table :<%= table_name %>, bulk: true do |t|
 <%= migration_data -%>
     end
   end