+++ lib/pay/stripe/billable.rb
@@ -26,23 +26,22 @@ module Pay
       #
       # Returns Stripe::Customer
       def customer
-        if processor_id?
+        stripe_customer = if processor_id?
           ::Stripe::Customer.retrieve(processor_id)
         else
-          stripe_customer = ::Stripe::Customer.create(email: email, name: customer_name)
-          billable.update(processor: :stripe, processor_id: stripe_customer.id)
-
-          # Update the user's card on file if a token was passed in
-          if card_token.present?
-            payment_method = ::Stripe::PaymentMethod.attach(card_token, {customer: stripe_customer.id})
-            stripe_customer.invoice_settings.default_payment_method = payment_method.id
-            stripe_customer.save
-
-            update_card_on_file ::Stripe::PaymentMethod.retrieve(card_token).card
-          end
-
-          stripe_customer
+          sc = ::Stripe::Customer.create(email: email, name: customer_name)
+          billable.update(processor: :stripe, processor_id: sc.id)
+          sc
         end
+
+        # Update the user's card on file if a token was passed in
+        if card_token.present?
+          payment_method = ::Stripe::PaymentMethod.attach(card_token, customer: stripe_customer.id)
+          stripe_customer = ::Stripe::Customer.update(stripe_customer.id, invoice_settings: {default_payment_method: payment_method.id})
+          update_card_on_file(payment_method.card)
+        end
+
+        stripe_customer
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
       end
@@ -85,6 +84,8 @@ module Pay
 
         # Inherit trial from plan unless trial override was specified
         opts[:trial_from_plan] = true unless opts[:trial_period_days]
+
+        # Load the Stripe customer to verify it exists and update card if needed
         opts[:customer] = customer.id
 
         stripe_sub = ::Stripe::Subscription.create(opts, {idempotency_key: idempotency_key})
@@ -203,8 +204,8 @@ module Pay
           payment_method_types: ["card"],
           mode: "payment",
           # These placeholder URLs will be replaced in a following step.
-          success_url: root_url,
-          cancel_url: root_url
+          success_url: options.delete(:success_url) || root_url,
+          cancel_url: options.delete(:cancel_url) || root_url
         }
 
         # Line items are optional
@@ -226,10 +227,11 @@ module Pay
       # checkout_charge(amount: 15_00, name: "T-shirt", quantity: 2)
       #
       def checkout_charge(amount:, name:, quantity: 1, **options)
+        currency = options.delete(:currency) || "usd"
         checkout(
           line_items: {
             price_data: {
-              currency: options[:currency] || "usd",
+              currency: currency,
               product_data: {name: name},
               unit_amount: amount
             },
@@ -242,7 +244,7 @@ module Pay
       def billing_portal(**options)
         args = {
           customer: processor_id,
-          return_url: options[:return_url] || root_url
+          return_url: options.delete(:return_url) || root_url
         }
         ::Stripe::BillingPortal::Session.create(args.merge(options))
       end