+++ lib/pay/stripe/error.rb
@@ -1,9 +1,7 @@
 module Pay
   module Stripe
     class Error < Pay::Error
-      def message
-        I18n.t("errors.stripe.#{result.code}", default: result.message)
-      end
+      delegate :message, to: :cause
     end
   end
 end