+++ test/pay/billable/braintree_test.rb
@@ -29,7 +29,6 @@ class Pay::Braintree::Billable::Test < ActiveSupport::TestCase
   end
 
   test "can update card" do
-    @billable.customer # Make sure we have a customer object
     @billable.update_card("fake-valid-discover-nonce")
     assert_equal "Discover", @billable.card_type
     assert_nil @billable.card_token
@@ -105,10 +104,20 @@ class Pay::Braintree::Billable::Test < ActiveSupport::TestCase
     @billable.card_token = "fake-processor-declined-visa-nonce"
     err = assert_raises(Pay::Braintree::Error) { @billable.subscribe }
     assert_equal "Do Not Honor", err.message
+    assert_equal Braintree::ErrorResult, err.cause.class
   end
 
   test "handles invalid parameters" do
     err = assert_raises(Pay::Braintree::AuthorizationError) { @billable.charge(10_00, metadata: {}) }
     assert_equal "Either the data you submitted is malformed and does not match the API or the API key you used may not be authorized to perform this action.", err.message
+  end
+
+  test "braintree card is automatically updated on subscribe" do
+    assert_nil @billable.card_type
+    @billable.update_card "fake-valid-discover-nonce"
+    assert_equal "Discover", @billable.card_type
+    @billable.card_token = "fake-valid-visa-nonce"
+    @billable.subscribe
+    assert_equal "Visa", @billable.card_type
   end
 end